CPLP-1077 move delete document to administration
