+++ src/registration/CatenaX.NetworkServices.Registration.Service/BusinessLogic/RegistrationBusinessLogic.cs
@@ -98,34 +98,6 @@ namespace CatenaX.NetworkServices.Registration.Service.BusinessLogic
             return await _portalDBAccess.SaveAsync().ConfigureAwait(false);
         }
 
-        /// <inheritdoc />
-        public async Task<bool> DeleteDocumentAsync(Guid documentId, string iamUserId)
-        {
-            var documentRepository = _portalRepositories.GetInstance<IDocumentRepository>();
-            var userRepository = _portalRepositories.GetInstance<IUserRepository>();
-            var companyUserId = await userRepository.GetCompanyUserIdForIamUserUntrackedAsync(iamUserId);
-            var document = await documentRepository.GetDocumentByIdAsync(documentId);
-
-            if (document is null)
-            {
-                throw new NotFoundException("Document is not existing");
-            }
-
-            if (document.CompanyUserId != companyUserId)
-            {
-                throw new ForbiddenException("User is not allowed to delete this document");
-            }
-
-            if (document.DocumentStatusId != DocumentStatusId.PENDING)
-            {
-                throw new ArgumentException("Incorrect document status");
-            }
-
-            await documentRepository.DeleteDocumentAsync(document);
-            await this._portalRepositories.SaveAsync().ConfigureAwait(false);
-            return true;
-        }
-
         public async IAsyncEnumerable<CompanyApplication> GetAllApplicationsForUserWithStatus(string userId)
         {
             await foreach (var applicationWithStatus in _portalDBAccess.GetApplicationsWithStatusUntrackedAsync(userId).ConfigureAwait(false))
